Move add_request_subscription
