+++ source/request-response/request_response_client.c
@@ -136,22 +136,6 @@ state, removed on operation completion/destruction
 
 */
 
-static struct aws_rr_response_path_entry *s_aws_rr_response_path_entry_new(
-    struct aws_allocator *allocator,
-    struct aws_byte_cursor topic,
-    struct aws_byte_cursor correlation_token_json_path) {
-    struct aws_rr_response_path_entry *entry = aws_mem_calloc(allocator, 1, sizeof(struct aws_rr_response_path_entry));
-
-    entry->allocator = allocator;
-    entry->ref_count = 1;
-    aws_byte_buf_init_copy_from_cursor(&entry->topic, allocator, topic);
-    entry->topic_cursor = aws_byte_cursor_from_buf(&entry->topic);
-
-    aws_byte_buf_init_copy_from_cursor(&entry->correlation_token_json_path, allocator, correlation_token_json_path);
-
-    return entry;
-}
-
 struct aws_mqtt_rr_client_operation {
     struct aws_allocator *allocator;
 
@@ -1202,30 +1186,7 @@ static int s_add_request_operation_to_response_path_table(
     struct aws_mqtt_rr_client_operation *operation) {
 
     struct aws_array_list *paths = &operation->storage.request_storage.operation_response_paths;
-    size_t path_count = aws_array_list_length(paths);
-    for (size_t i = 0; i < path_count; ++i) {
-        struct aws_mqtt_request_operation_response_path path;
-        aws_array_list_get_at(paths, &path, i);
-
-        struct aws_hash_element *element = NULL;
-        if (aws_hash_table_find(&client->subscriptions.request_response_paths, &path.topic, &element)) {
-            return aws_raise_error(AWS_ERROR_MQTT_REQUEST_RESPONSE_INTERNAL_ERROR);
-        }
-
-        if (element != NULL) {
-            struct aws_rr_response_path_entry *entry = element->value;
-            ++entry->ref_count;
-            continue;
-        }
-
-        struct aws_rr_response_path_entry *entry =
-            s_aws_rr_response_path_entry_new(client->allocator, path.topic, path.correlation_token_json_path);
-        if (aws_hash_table_put(&client->subscriptions.request_response_paths, &entry->topic_cursor, entry, NULL)) {
-            return aws_raise_error(AWS_ERROR_MQTT_REQUEST_RESPONSE_INTERNAL_ERROR);
-        }
-    }
-
-    return AWS_OP_SUCCESS;
+    return aws_mqtt_request_response_client_subscriptions_add_request_subscription(&client->subscriptions, paths);
 }
 
 static int s_add_request_operation_to_correlation_token_table(
